adicionando controller DispositivoIotResource

+++ src/main/java/br/com/fiap/esgsmart/repository/DispositivoIotRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface DispositivoIotRepository extends JpaRepository<DispositivoIoT, Integer> {
+public interface DispositivoIotRepository extends JpaRepository<DispositivoIoT, Long> {
 
 
     List<DispositivoIoT> findByLocalizacao(String localizacao);
